+++ cloudmesh/provider/command/provider.py
@@ -1,7 +1,10 @@
+import json
 import os
+import textwrap
 from pydoc import locate
 
 from cloudmesh.common.Tabulate import Printer
+from cloudmesh.common.console import Console
 from cloudmesh.shell.command import PluginCommand
 from cloudmesh.shell.command import command, map_parameters
 
@@ -19,6 +22,7 @@ class ProviderCommand(PluginCommand):
 
            Usage:
              provider list [--output=OUTPUT]
+             provider info SERVICE NAME WHAT
 
            Arguments:
              NAME           The name of the key.
@@ -29,15 +33,19 @@ class ProviderCommand(PluginCommand):
 
            Description:
 
-                THIS IS NOT YET IMPLEMENTED
+                What: output, sample
 
-                Managing the providers
+           Examples:
+             Getting the sample and output from provides via a command
+
+               cms provider info compute openstack sample
+               cms provider info compute openstack output
+               cms provider list --output=json
+               cms provider list
+
         """
 
-        map_parameters(arguments, 'output')
-
-        if arguments.list:
-
+        def craete_list():
             _paths = []
 
             def find(names,
@@ -46,18 +54,25 @@ class ProviderCommand(PluginCommand):
                      on_error="not loaded"):
                 _paths = []
                 for _name in names:
+
                     _active = False
                     try:
-                        _where = os.path.dirname(
-                            locate(template.format(name=_name)).__file__)
+                        _class = template.format(name=_name)
+                        _p = locate(_class)
+                        _where = _p.__file__
+                        _path = os.path.dirname(_where)
                         _active = True
+                        _provider = locate(f"{_class}.Provider.Provider")
                     except Exception as e:
-                        _where = on_error.format(name=_name)
+                        _path = on_error.format(name=_name)
+                        _provider = None
                     _paths.append({
-                        "path": _where,
+                        "class": _class,
+                        "path": _path,
                         "name": _name,
                         "active": _active,
-                        "kind": kind
+                        "service": kind,
+                        "provider": _provider
                     })
                 return _paths
 
@@ -82,7 +97,7 @@ class ProviderCommand(PluginCommand):
                           kind="compute",
                           on_error="load with pip install cloudmesh-cloud") + \
                      find(candidates_name_compute,
-                          template="cloudmesh.{_name}.compute",
+                          template="cloudmesh.{name}.compute",
                           kind="compute",
                           on_error="load with pip install cloudmesh-{name}")
 
@@ -108,7 +123,7 @@ class ProviderCommand(PluginCommand):
                            kind="storage",
                            on_error="load with pip install cloudmesh-storage") + \
                       find(candidates_name_storage,
-                           template="cloudmesh.{_name}.storage",
+                           template="cloudmesh.{name}.storage",
                            kind="storage",
                            on_error="load with pip install cloudmesh-{name}")
 
@@ -117,15 +132,62 @@ class ProviderCommand(PluginCommand):
                 "azure"
                 "google",
                 "multipass",
-                "opensatck"
+                "opensatck",
                 "oracle"]
 
             _paths += find(candidates_volume,
                            template="cloudmesh.volume.provider.{name}",
                            kind="volume",
                            on_error="load with pip install cloudmesh-volume")
+            return _paths
 
+        map_parameters(arguments, 'output')
+
+        if arguments.info:
+            try:
+                service = arguments.SERVICE
+                name = arguments.NAME
+                what = arguments.WHAT
+
+                services = craete_list()
+
+                for provider in services:
+
+                    try:
+
+                        if provider['service'] == service and \
+                            provider['name'] == name:
+
+                            if arguments.WHAT == 'sample':
+                                print(textwrap.dedent(
+                                    provider["provider"].sample))
+                            elif arguments.WHAT == 'output':
+                                print(json.dumps(provider["provider"].output,
+                                                 indent=4))
+                                print()
+
+
+                    except Exception as e:
+                        print(e)
+
+
+
+
+            except:
+                Console.error("Problem getting the Provider info")
+                return ""
+
+        elif arguments.list:
+
+            print(arguments.output)
+            _paths = craete_list()
+
+            for entry in _paths:
+                del entry["provider"]  # can not be printed
             print(
-                Printer.write(_paths, order=["kind", "name", "active", "path"]))
+                Printer.write(_paths,
+                              order=["service", "name", "active", "path"],
+                              output=arguments.output)
+            )
 
         return ""